Struct rasn::ber::de::Decoder

pub struct Decoder<'input> { /* private fields */ }

A BER and variants decoder. Capable of decoding BER, CER, and DER.

Implementations

impl<'input> Decoder<'input>

pub fn new(input: &'input [u8], config: DecoderOptions) -> Self

Create a new Decoder from the given input and config.

pub fn decoded_len(&self) -> usize

Return a number of the decoded bytes by this decoder