Module rasn::types

ASN.1 Data Types

The types modules is a collection of Rust types and data structures that are defined to represent various ASN.1 data types, and renamed to use ASN.1’s terminology.

Structs

Any

Represents a complete encoded ASN.1 value of any type. Usually identified with an ObjectIdentifier.

ConstOid

A temporary workaround for Oid not currently being const compatible.

Explicit

A newtype wrapper that will explicitly tag its value with T’s tag.

Implicit

A newtype wrapper that will implicitly tag its value with T’s tag.

InstanceOf

An instance of a defined object class.

Integer

A big signed integer type.

ObjectIdentifier

A global unique identifier that identifies an concept, such as a organisation, or encoding rules. The “owned” version of Oid.

OctetString

A cheaply cloneable and sliceable chunk of contiguous memory.

Oid

A reference to a global unique identifier that identifies an concept, such as a organisation, or encoding rules.

Tag

An abstract representation of an ASN.1 tag that uniquely identifies a type within a ASN.1 module for codecs.

Utf8String

A UTF-8–encoded, growable string.

Enums

Class

The class of tag identifying its category.

Open

An “open” type representing any valid ASN.1 type.

TagTree

The root or node in tree reprensenting all of potential tags in a ASN.1 type. For most types this is only ever one level deep, except for CHOICE enums which will contain a set of nodes, that either point to a Leaf or another level of Choice.

Traits

AsnType

A trait representing any type that can represented in ASN.1.

Type Definitions

BitString

The BIT STRING type.

BmpString

The BmpString type.

GeneralString

The GeneralString type.

GeneralizedTime

The GeneralizedTime type.

Ia5String

The Ia5String type.

NumericString

The NumericString type.

PrintableString

The PrintableString type.

SequenceOf

The SEQUENCE OF type.

SetOf

The SET OF type.

TeletexString

The TeletexString type.

UniversalString

The UniversalString type.

UtcTime

The UTCTime type.

VisibleString

The VisibleString type.

Derive Macros

AsnType

An automatic derive of the AsnType trait.